+++ NEWS.md
@@ -1,6 +1,10 @@
 # tidycmprsk (development version)
 
-* Update the `crr()` print method for more clarity.
+* Now returning the the number at risk, number of events, and the number of censored observations in `tidy.tidycuminc()`.
+
+* Updated the default value to `tidy.tidycuminc(conf.int = TRUE)` from `FALSE`
+
+* Update the `crr()` and `cuminc()` print methods for more clarity.
 
 # tidycmprsk 0.1.0